refatcor: separate Map from Home page
