+++ src/assets/styles/home.scss
@@ -18,6 +18,7 @@ body {
   height: 100%;
   position: relative;
 }
+
 .map__container {
   width: 100%;
   height: 100%;
