+++ src/components/Footer.jsx
@@ -1,0 +1,9 @@
+import { Route } from 'react-router-dom'
+
+export default function Footer (props) {
+
+
+  return (
+    
+  )
+}
